remote config: Create a cached and a no cached version of the scenarios (#424)
Whether or not a tracer caches config files is an opt in feature. If a
tracer does opt-in to caching the script needs to change to only send
new or modified config files to the tracer so that the tracer is
properly challenged to detect and verify unmodified files by using its
cache. Since everything is static as far as the scripts go, this means
we need a CACHE and a NOCACHE scenario for each product.