remote_config: append /0.7/config to info endpoint in rc scenario (#454)
When tracers start up, they query the /info endpoint of the agent to
determine what features the agent is capable of. In the remote config
scenarios, at this time we are not enabling remote config in the agent
since we are capturing requests to /0.7/config in the proxy and usingpre
mocked responses. Thus, /0.7/config is missing in the list of available
endpoints and production tracers won't enable the remote config client.

To fix this, we also modify /info when running the remote config
scenario to append this endpoint to the list so that tracers will work
properly.

Co-authored-by: Charles de Beauchesne <charles.de.beauchesne@gmail.com>