Add Java library later in the Docker build process
Building a Java application is slow (fetching dependencies, compiling),
while adding the library jar to the Docker image is fast, and possibly
more frequent.

By moving the step of adding the library jar later in the process, we
can still reuse most of the Docker build cache after changing to a
different library jar, which is particularly useful for local testing.