apps/golang: fix flush and enable span sampling tests
Enable single span sampling tests for the Go Tracer after fixes in https://github.com/DataDog/dd-trace-go/pull/1509.

Some bugs in the test setup needed to be fixed in order to accomplish this:
  - The Flush APIs needed to call Flush instead of Stop
  - One of the tests was flaky at 1 second. Moving it to 2 appears to deflake this.

This PR updates the go.mod to point to https://github.com/DataDog/dd-trace-go/pull/1509, which is currently under review.