Add format.sh for easy black formatting (#682)
Since setting up a virtualenv might be a hassle for some, and some users
have reported difficulties to get it installed correctly in their systems,
the script `format.sh` adds a simple way to run a dockerized black. The
only requisite is having Docker installed, and a Unix-like host (macOS
or Linux).