[RCM] Add test ensuring tracers don't hardcode config name
The last component of a config's path is the "name". RC was built with
the intention of allowing multiple config's to be associated with a
given config ID, but in practice thus far it has always been a single
config that the backend hard-codes to "config". This may change, so
tracers should NOT assume that "config" is the only possible config
name.