Add a basic `shell.nix` file
This is entirely optional to use but for nixpkgs users this is a really
easy way to get things up and running. With just:

```
nix-shell
```

All needed dependencies get automatically pulled in, dropping the
user in a ready-to-use shell.

Then as usual:

```
pip install -r requirements.txt
```

Puts everything in a `vendor/python` directory, which behaves like a
virtualenv. Should it need to be cleaned up, e.g to start form scratch,
a simple `rm -rf vendor/python` is sufficient.

It also pins to a specific commit for the channel, so that tool versions
are always consistent.