Fix /etc/otelcol-config.yml perms in Linux
Depending on the umask settings, the otel conf path might be set without
read permissions by other users. Since we bind mount this file into a
Docker container, and the user within the container is different from
the one in the host, the file requires permission to read by other
users. Force it when needed before starting the container.