Test the test: Validate json report generation
Incorrect generation of json report could cause problems in system-tests-dashboard due to we are parsing the report.
This PR adds a new test to validate it.
To generate the report we have to launch a new scenario "mock_the_test".This scenario is launched programatically executing ./run.sh MOCK_THE_TEST