Fix pretty asserts in utilities files (#1459)
Pytests transforms tests to convert calls to `assert` to richer pretty
asserts. However, it does so only in files scanner for tests. That is,
those matched by the `test_*.py` pattern.

This change adds `_test_*.py` to the list of valid test file patterns,
and moves existing test utilities to use that convention and benefit
from pretty asserts.

Pretty assertions include the values of variables used in assertions,
diffs, etc.