Add decode support for capability byte arrays
Turns out due to an implementation detail in the go json library, byte
arrays can be encoded equally as base64 or an array of numbers.

The Python, .NET and Node libraries (at least) use base64 encoding to
encode the capabilities and it seems like the Java implementation uses
an array of numbers.

Since both are currently accepted by the agent, both should be accepted
by the tests.

In the future we should standardize to avoid further unexpected behaviour.