Update configuration tests with JSON types
We're permitting that types be sent as JSON in addition to strings so
that the backend can normalize values like `null` consistently.

Other changes include supporting "inactive" for appsec_enabled which is
supported in Java and indicates that the library is currently disabled
but can be enabled using Remote Config.