go: add Go test app for profiling/tracing "SSI"
Go doesn't use "SSI", i.e. run-time instrumentation, but it does have
automatic instrumentation at compile time via Orchestrion. Add tests for
that similar to the SSI one for other languages, where we run an
instrumented HTTP server application and check that we generate traces
and profiles and that they make it through our full processing pipeline.

We need some kludges to make sure this stuff works for Go. We don't have
"golang" injection for the agent installer, and the script fails if we
try to run it, so we tell the installer script we're "java" just to make
it complete.