[Java] Weblog endpoint prompt rules (#4743)
Motivation
In Java, we need to support at least seven different implementations of the same endpoint to cover all the weblog variants we support.

This means that whenever a new test is created that affects Java, we need to implement all the endpoints that the test needs for each of these variants.

These rules are intended to streamline and automate the creation of those endpoints.

Changes
Nested cursor rule and documentation added

Additional Notes
It has been tested to generate the endpoints for tests/appsec/iast/test_sampling_by_route_method_count.py with good results, although it usually requires some iteration to correctly generate the endpoint logic.