Disable Client-Side-Stats on default tests
dd-trace-go is about to enable Client-Side-Stats by default. This
feature causes all P0 traces to be dropped on the tracer side.

Many system-tests targeting all tracers rely on the fact that all traces
are flushed to the agent, and to avoid breaking this pattern, the FF
controlling CSS on Go is set to false.

A follow up PR(s) will enhance CSS coverage, which is currently rather
low.