[NO-TICKET] Enable crashtracking test for Ruby
**What does this PR do?**

This PR enables the crashtracking test for Ruby. There isn't
a lot of detail on https://datadoghq.atlassian.net/browse/APMLP-335
for why it was disabled, but I suspect
https://github.com/DataDog/system-tests/pull/4804 was the reason.

Now that #4804 is fixed, we should be able to enable the test.

**Motivation:**

Validate that crashtracking is in good shape for Ruby!

**Additional Notes:**

N/A

**How to test the change?**

Validate that the test correctly passes.