tests/debugger: allow error for "received" case in Go
In the other languages, if you can't find a method, then it might exist
but hasn't been loaded yet. For Go, you definitively know whether it
exists and can tell the user. Because we can, we do. The test needs to
be adjusted to permit this case. We track the history and assert that
indeed we *also* got the received diagnostic.