[Java] Temporarily disable oss.sonatype.org to unblock Java builds (#5017)
Motivation
The repository oss.sonatype.org is currently under maintenance, which is causing issues when downloading dependencies and breaking Java builds.

As a temporary workaround—and while we look for a more stable solution—the repository has been commented out to unblock the system-tests builds.

For reference, this is the original PR that introduced the repository:
#4310