feat(di): Tracers should write to debugger/v2/input when available
We are working on redirecting snapshots from the logs track to the debugger intake. We are doing this with a fallback for backwards compatibility with the agent.

The fallback works like so:

1. Write the snapshots to the `/debugger/v2/input` endpoint when available, otherwise
2. Fallback and write snapshots to the `/debugger/v1/diagnostics` endpoint.

These two tests `test_log_line_snapshot_debug_track` and `test_log_line_snapshot_new_destination` should cover both of these cases.

Refs: DEBUG-4420