[Java] Change missing_feature with irrelevant spring-boot-3-native (#5344)
Motivation
The spring-boot-3-native weblog variant is currently marked as a missing_feature in most tests. This causes the tests to still run for that variant, even though we’re not planning to support it anytime soon. Marking it as irrelevant instead will prevent those tests from running, helping to reduce unnecessary execution and lighten the load on the pipelines.

Changes
Change missing_feature with irrelevant spring-boot-3-native