DEBUG-4541: relax schema for Go symbols
Don't insist for "env" field to be present on SymDB uploads. This field
is not read by the backend, and is not populated on Go uploads. The
presence of this field from the other tracers is confusing, since it
suggests that an upload is associated with an env when it isn't; it's
only associated with a service name / version.