Disable Ruby 2.7-3.3 for multicontainer
This is a mitigation for crossing the 1h timeout.

These versions should be extracted to single container cases, while
multicontainer should test an actually interesting multicontainer.

Meanwhile, we keep the min and max, as well as the commented out bits
and dead Dockerfiles for them to be moved sooner rather than later.