prepare test for future go library
The Go library is learning new capabilities. This patch adds the
upcoming expected capabilities and disables the test for the current
version of the tracer: the released version of the tracer is 2.4.0; the
main branch advertises the current version as 2.6.1. So, I'm skipping
the test for versions <=2.6.1 until I merge my tracer changes, then I'll
only skip for <= 2.4.0.